+++ src/main/java/org/kinscript/Academy/dominio/dto/CursosDto.java
@@ -3,7 +3,7 @@ package org.kinscript.Academy.dominio.dto;
 import jakarta.validation.constraints.NotBlank;
 
 public record CursosDto (
-    Integer idCurso,
+        Long idCurso,
     @NotBlank(message = "El nombre del curso es obligatorio")
     String nombreCurso
 ){
