+++ src/main/java/org/kinscript/Academy/dominio/dto/JornadasDto.java
@@ -3,7 +3,7 @@ package org.kinscript.Academy.dominio.dto;
 import jakarta.validation.constraints.NotBlank;
 
 public record JornadasDto(
-        Integer idJornada,
+        Long idJornada,
         @NotBlank(message = "El nombre de la jornada es obligatorio")
         String nombreJornada
 ) {
