+++ src/main/java/org/kinscript/Academy/dominio/exception/GradoCursoNotExistsException.java
@@ -1,7 +1,7 @@
 package org.kinscript.Academy.dominio.exception;
 
 public class GradoCursoNotExistsException extends RuntimeException {
-    public GradoCursoNotExistsException(Integer idGrado, Integer idCurso) {
+    public GradoCursoNotExistsException(Long idGrado, Long idCurso) {
         super("No existe una asociación para el grado ID " + idGrado + " y el curso ID " + idCurso + ".");
     }
 }
