+++ src/main/java/org/kinscript/Academy/dominio/exception/ProfesorAsignacionExistsException.java
@@ -1,7 +1,7 @@
 package org.kinscript.Academy.dominio.exception;
 
 public class ProfesorAsignacionExistsException extends RuntimeException {
-    public ProfesorAsignacionExistsException(Integer idProfesor, Integer idCurso, Integer idGrado, Integer idSeccion) {
+    public ProfesorAsignacionExistsException(Long idProfesor, Long idCurso, Long idGrado, Long idSeccion) {
         super("La asignación para el profesor " + idProfesor + ", curso " + idCurso +
                 ", grado " + idGrado + " y sección " + idSeccion + " ya existe.");
     }
